hf_pipeline/code_gen/requirements.txt
# Code Generator Requirements

# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
gradio>=4.0.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.0.0

# Optional: For better performance
# torch>=2.0.0
# transformers>=4.30.0
